fork bomb, creates a zillion child processes , DO NOT RUN THIS CODE on your System (creates DOS)
As you may suspect, the code creates zillions of children in order to overpower the OS, and if the OS follows the code and does what has been ordered, it will die in torment after a few seconds.

Remember, you have been warned – it is a bomb. It may explode.